Fixed-return-type intrinsics return Long, not the magnitude rule (#193)
Len, InStr, InStrRev, LBound, UBound and VarType returned Integer whenever the
answer fitted Int16. VB6 returns Long from all of them regardless of magnitude.

None of them chose a type. They built their result from a C# int, and
Vb6Value(int) applies a magnitude rule: anything fitting Int16 is reported as an
Integer. That rule is right for an arithmetic literal, whose type genuinely does
follow its magnitude, and wrong for a function with a declared return type.
Vb6Value(long) has no such rule, so the fix is to construct these as Long.

Measured the whole family first rather than fixing the five in the issue, and
that turned up two the issue did not have. VarType is LONG despite every vbXxx
code fitting an Integer, and the comment beside it asserted Integer. DateDiff is
Long while DatePart is Integer — a pair that looks like an inconsistency to be
tidied and is not.

It is not one rule applied everywhere, which is why a blanket widening would
have been just as wrong in the other direction: Asc, Sgn, Year, Month, Day,
Hour, Minute, Second, Weekday and DatePart really are Integer. All ten are now
tests. Int and Fix are left alone deliberately — they preserve their operand's
subtype rather than having a fixed one, so Int(3) is Integer and Int(3.5) is
Double, and they must keep going through the magnitude rule.

Twenty-one existing assertions across six test classes pinned the old Integer.
Every one encoded HexIDE's bug rather than VB6's behaviour, so they are
corrected rather than worked around — including the #190 test that deliberately
asserted the wrong value with a comment pointing here. That pointer did its job:
the divergence was recorded at the exact spot someone would otherwise have
"corrected" the test to match the bug.

Asserted through TypeName rather than by value, because the value was never
wrong — only the subtype, which is invisible until something asks. It does not
stay invisible: it feeds the arithmetic result-type ladder, so Len("hi") + 1 is
a Long now, and that is a test too.

Full suite green at 1191.

+++ IDE/HexIDE.Runtime/Interpreter/VB6BuiltIns.Inspection.cs
@@ -16,7 +16,9 @@ public partial class VB6BuiltIns
     private static void RegisterInspection(Dictionary<string, BuiltinFn> d)
     {
         d["TypeName"]  = (_, a, _) => new Vb6Value(TypeNameOf(a[0]));
-        d["VarType"]   = (_, a, _) => new Vb6Value(TypeInfo(a[0]).code);      // vbXxx code, an Integer
+        // Measured: VarType is LONG, despite every vbXxx code fitting an Integer — the comment here used to
+        // say Integer and was wrong. Fixed declared return type, so not the magnitude rule. (#193)
+        d["VarType"]   = (_, a, _) => new Vb6Value((long)TypeInfo(a[0]).code);
         d["IsNumeric"] = (_, a, _) => new Vb6Value(IsNumericValue(a[0]));
         d["IsDate"]    = (_, a, _) => new Vb6Value(IsDateValue(a[0]));
         d["IsEmpty"]   = (_, a, _) => new Vb6Value(a[0].Type == VT.EmptyVariant);
